dags/requirements.txt
azure-storage-file-datalake

Pillow
apache-airflow-providers-microsoft-azure
